Python/requirements.txt
chardet==4.0.0
configparser==5.0.1
mysql-connector==2.2.9
pdfminer.six==20200517
pdfplumber==0.5.25
Pillow==8.0.1
pycryptodome==3.9.9
sortedcontainers==2.3.0
uuid==1.30
Wand==0.6.5